+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Features:
 - [BDN-990](https://appodeal.atlassian.net/browse/BDN-990) Updated compileSdkVersion to 35
 - [BDN-954](https://appodeal.atlassian.net/browse/BDN-954) Migrated to build-logic
+- [BDN-944](https://appodeal.atlassian.net/browse/BDN-944) Use endpoint for BidMachine configuration
 
 # 0.9.0 (2025.07.24)
 ## Features: